Calculating population-averaged years of life lost

Calculate the population-averaged years of life lost due to radiation exposure.

Usage

population_YLL(dsGy, reference, riskmodel, agex = 1:8 * 10 - 5, nmc = 10000)

Arguments

dsGy

radiation dose in Gy or Sv (a single value).

reference

baseline rate, all cause mortality rate and age distribution in the reference population (a list object, which contains data.frame objects named 'baseline' for baseline rates of the target endpoint, 'mortality' for all cause mortality rates and 'agedist' for age distribution in the reference population).

riskmodel

risk model risk model (a list object, which contains two list objects for excess relative risk model (err) and excess absolute risk model (ear), each of which contains a vector of parameter values (para), a matrix of variance covariance matrix (var), and a function to compute the risk given a parameter vector, a dose value, an age at exposure, an attained age and sex.

agex

a vector of ages at exposure, which represent age categories (default values: 5, 15, 25, ..., 75 to represent age categories 0-10, 10-20, ..., 70-80) option for risk calculation (a list object, which contains maximum age to follow up (an integer value)

mmc

an integer for the Monte Carlo sample size (default: 10000)

Value

estimated risk information (list)

Examples

 # The following examples use default data provided in CanEpiRisk package
 # for riskmodels (LSS_mortality) derived from Life Span Study
 #     baseline rates and age distribution for WHO riskmodels (Mortality, agedist_rgn)

 # Example: allsolid mortality, Region-1, exposed to 0.1 Gy, followed up to age 100, LSS linear ERR
 ref1 <- list(  baseline=Mortality[[1]]$allsolid,     # baseline rates
               mortality=Mortality[[1]]$allcause,     # allcause mortality
                 agedist=agedist_rgn[[1]] )           # age distribution
 mod1 <- LSS_mortality$allsolid$L                     # risk model
 population_YLL( dsGy=0.1, reference=ref1, riskmodel=mod1 )    # YLL
#> $err
#>            male     male_lo    male_up      female   female_lo  female_up
#> 5   0.172548454 0.113541561 0.25203103 0.344979390 0.243108697 0.49008060
#> 15  0.120426740 0.083455356 0.16511131 0.241120226 0.186880910 0.30829843
#> 25  0.084265120 0.058227720 0.11586864 0.167816142 0.134216229 0.20817007
#> 35  0.058837161 0.037990152 0.08649037 0.112293414 0.084437374 0.14922384
#> 45  0.038807930 0.022672157 0.06342415 0.069751324 0.046986903 0.10362435
#> 55  0.021701822 0.011381537 0.03980908 0.037673823 0.022324610 0.06334045
#> 65  0.009428082 0.004390760 0.01961776 0.016326514 0.008465679 0.03156331
#> 75  0.002930897 0.001207534 0.00696487 0.004866801 0.002206870 0.01075800
#> all 0.066209310 0.046037923 0.09004550 0.121945509 0.096877403 0.15157949
#>             all      all_lo      all_up
#> 5   0.257192418 0.188559222 0.351876583
#> 15  0.179734592 0.145378805 0.220969519
#> 25  0.125319482 0.103636149 0.151529142
#> 35  0.085316148 0.065255003 0.111621195
#> 45  0.054564407 0.036594333 0.080564019
#> 55  0.029898221 0.017555962 0.050533298
#> 65  0.013118623 0.006732740 0.025384365
#> 75  0.004027477 0.001795772 0.008936945
#> all 0.094648947 0.078545676 0.113775705
#> 
#> $ear
#>            male     male_lo    male_up      female   female_lo  female_up
#> 5   0.181753803 0.121733560 0.26421792 0.281421457 0.209593267 0.37510983
#> 15  0.145920191 0.102930962 0.20101727 0.226262775 0.183228213 0.27835336
#> 25  0.114559344 0.081410955 0.15745705 0.177872589 0.145531196 0.21606420
#> 35  0.084993194 0.057697876 0.12230769 0.132968146 0.102510151 0.17186563
#> 45  0.057735298 0.036355773 0.08940376 0.091547990 0.064385510 0.13060683
#> 55  0.034196841 0.019704211 0.05813483 0.055469841 0.035274583 0.08858777
#> 65  0.016747634 0.008787176 0.03188115 0.027640334 0.015785079 0.04990980
#> 75  0.005890745 0.002780126 0.01248911 0.009799129 0.004971703 0.01996568
#> all 0.084027402 0.059662849 0.11561861 0.124987530 0.102390886 0.15167382
#>             all      all_lo     all_up
#> 5   0.230588887 0.175450422 0.30469752
#> 15  0.185633721 0.153531314 0.22479636
#> 25  0.145745417 0.121613153 0.17449858
#> 35  0.109179689 0.085045809 0.13977903
#> 45  0.074964804 0.053128661 0.10647285
#> 55  0.045278843 0.028830563 0.07200907
#> 65  0.022708162 0.012852234 0.04068785
#> 75  0.008134338 0.004114722 0.01650545
#> all 0.104984503 0.087905496 0.12561305
#>