Calculating cumulative excess risks

Calculate the cumulative excess risk due to radiation exposure.

Usage

CER(exposure, reference, riskmodel, option)

Arguments

exposure

a list object that specifies the exposure scenario, which contains agex (a single value or a vector for age(s) at exposure), doseGy (a single value or a vector of dose(s) in Gy), and sex (1 or 2 for male or female).

reference

a list object that specifies the baseline information of the reference population, which contains data.frame objects named baseline for baseline rates of the target endpoint and 'mortality' for all cause mortality rates.

riskmodel

a list object that specifies the risk model, which contains two list objects named err for excess relative rate model and ear for excess absolute rate model, each of which contains a vector para for model parameter estimates, a matrix var for the variance covariance matrix, and a function f to compute the excess risk given a parameter vector and exposure information (e.g., dose, age at exposure, sex, attained age).

option

a list object that specifies optional settings for risk calculation, which contains an integer value maxage for the maximum age to follow up, a value err_wgt for the weight for risk transfer (1=err, 0=ear), an integer value n_mcsamp for the number of Monte Carlo samples, and alpha for the significance level (default=0.05).

Value

information of calculated risk (vector)

See also

Examples

 # The following examples use default data provided in CanEpiRisk package
 # for riskmodels (LSS_mortality and LSS_incidence) derived from Life Span Study
 # and baseline mortality and incidence rates for WHO global regions (Mortality and Incidence).

 # Example 1: allsolid mortality, Region-1, female, 0.1Gy at age 15, followed up to age 100, LSS linear ERR
 exp1 <- list( agex=5, doseGy=0.1, sex=2 )   # exposure scenario
 ref1 <- list( baseline=Mortality[[1]]$allsolid,        # baseline rates
              mortality=Mortality[[1]]$allcause )       # all-cause mortality
 mod1 <- LSS_mortality$allsolid$L                       # risk model
 opt1 <- list( maxage=100, err_wgt=1, n_mcsamp=10000 )  # option
 CER(  exposure=exp1, reference=ref1, riskmodel=mod1, option=opt1 ) * 10000 # cases per 10,000
#>      mle     mean   median  ci.2.5% ci.97.5% 
#> 221.1383 227.2963 221.3405 147.4744 337.4809 

 # Example 2: leukaemia incidence, Region-4, male, 6.7(100/15)mGy at ages 30-45, followed up to age 60, LSS LQ EAR
 exp2 <- list( agex=30:44+0.5, doseGy=rep(0.1/15,15), sex=1 )
 ref2 <- list( baseline=Incidence[[4]]$leukaemia,       # baseline rates
              mortality=Mortality[[4]]$allcause )       # all-cause mortality rates
 mod2 <- LSS_incidence$leukaemia$LQ                     # risk model
 opt2 <- list( maxage=60, err_wgt=0, n_mcsamp=10000)    # option
 CER(  exposure=exp2, reference=ref2, riskmodel=mod2, option=opt2 ) * 10000 # cases per 10,000
#>         mle        mean      median     ci.2.5%    ci.97.5% 
#>  3.68374227  3.64300158  3.61730996 -0.04646003  7.29646319