Calculating population-averaged lifetime attributable risks

Calculate the population-averaged lifetime attributable risk due to radiation exposure.

Usage

population_LAR(
  dsGy,
  reference,
  riskmodel,
  agex = 1:8 * 10 - 5,
  PER = 10000,
  nmc = 10000
)

Arguments

dsGy

radiation dose in Gy or Sv (a single value).

reference

baseline rate, all cause mortality rate and age distribution in the reference population (a list object, which contains data.frame objects named baseline for baseline rates of the target endpoint, 'mortality' for all cause mortality rates and 'agedist' for age distribution in the reference population).

riskmodel

risk model risk model (a list object, which contains two list objects for excess relative risk model (err) and excess absolute risk model (ear), each of which contains a vector of parameter values (para), a matrix of variance covariance matrix (var), and a function to compute the risk given a parameter vector, a dose value, an age at exposure, an attained age and sex.

agex

a vector of ages at exposure, which represent age categories (default values: 5, 15, 25, ..., 75 to represent age categories 0-10, 10-20, ..., 70-80) option for risk calculation (a list object, which contains maximum age to follow up (an integer value)

PER

an integer (default value: 10000 to show the estimated risk as cases per 10000)

mmc

an integer for the Monte Carlo sample size (default: 10000)

Value

estimated risk information (list)

See also

Examples

 # The following examples use default data provided in CanEpiRisk package
 # for riskmodels (LSS_mortality and LSS_incidence) derived from Life Span Study
 # and baseline mortality and incidence rates for WHO global regions (Mortality and Incidence).

 # Example 1: allsolid mortality, Region-1, exposed to 0.1 Gy, followed up to age 100, LSS linear ERR
 ref1 <- list(  baseline=Mortality[[1]]$allsolid,     # baseline rates
               mortality=Mortality[[1]]$allcause,     # allcause mortality
                 agedist=agedist_rgn[[1]] )           # age distribution
 mod1 <- LSS_mortality$allsolid$L                     # risk model
 population_LAR( dsGy=0.1, reference=ref1, riskmodel=mod1 )    # CER cases per 10,000
#> $err
#>           male   male_lo   male_up     female  female_lo female_up        all
#> 5   129.985169 83.244545 202.37662 221.360084 147.407145 337.76134 175.041033
#> 15   92.016631 62.810077 131.80442 156.370934 115.809820 214.45519 123.749165
#> 25   65.551025 45.061006  91.27050 110.471441  85.913326 143.50363  87.690237
#> 35   46.616567 30.433884  68.02224  77.181659  57.753338 103.72140  61.950954
#> 45   32.670005 19.608545  52.92543  52.453651  35.309729  77.04837  42.782281
#> 55   21.289406 11.499543  38.58884  33.473888  19.977769  55.47146  27.592565
#> 65   12.299479  5.881530  25.08351  19.042264   9.956742  35.73864  15.907303
#> 75    5.826031  2.447322  13.43129   8.653533   3.944694  18.54279   7.432784
#> all  52.951243 36.438037  74.01889  83.720535  64.500929 110.00198  68.700499
#>         all_lo    all_up
#> 5   121.025569 258.73475
#> 15   95.532639 163.81352
#> 25   70.529820 110.35512
#> 35   47.082869  82.16893
#> 45   29.261592  62.84587
#> 55   16.512574  45.89406
#> 65    8.319541  30.25541
#> 75    3.394867  16.09819
#> all  55.082630  87.33942
#> 
#> $ear
#>          male   male_lo   male_up    female  female_lo female_up       all
#> 5   119.45689 77.149979 186.33184 177.54254 126.246809 254.36511 148.19290
#> 15   96.61196 65.573878 140.50387 143.55266 111.061348 186.51045 119.69381
#> 25   77.71118 53.913547 110.30544 115.45280  92.668914 143.59826  96.49369
#> 35   61.56552 41.415121  89.84424  91.50250  70.626980 117.63294  76.56459
#> 45   47.09130 29.586586  74.63232  70.15035  49.938165  99.23116  58.84359
#> 55   33.83488 19.450301  59.01992  50.76188  32.436397  80.22252  42.65354
#> 65   22.16773 11.542205  43.14237  33.34129  19.107819  59.42208  28.20988
#> 75   11.99442  5.593139  26.19668  18.12527   9.214657  36.28217  15.50287
#> all  61.30818 42.381433  87.84699  87.89673  69.899239 110.29007  74.98127
#>         all_lo    all_up
#> 5   106.084994 210.57351
#> 15   94.169928 154.82458
#> 25   78.306626 119.79742
#> 35   59.516218  99.24809
#> 45   41.775971  84.00931
#> 55   27.198562  67.93516
#> 65   16.061186  50.42921
#> 75    7.837466  31.34409
#> all  60.216621  94.17161
#> 

 # Example 2: leukaemia incidence, Region-4, exposed to 0.1 Gy, followed up to age 100, LSS LQ ERR
 ref2 <- list(  baseline=Incidence[[4]]$leukaemia,    # baseline rates
               mortality=Mortality[[4]]$allcause,     # all-cause mortality
                 agedist=agedist_rgn[[4]] )           # age distribution
 mod2 <- LSS_incidence$leukaemia$LQ                   # risk model
 population_LAR( dsGy=0.1, reference=ref2, riskmodel=mod2 )    # CER cases per 10,000
#> $err
#>          male     male_lo  male_up    female   female_lo female_up       all
#> 5   16.859485 -0.30151992 69.95874 12.735814 -0.22969726  51.71645 14.884020
#> 15   8.040584 -0.12615249 19.97858  6.494689 -0.10235328  15.64918  7.303086
#> 25   6.341666 -0.08306336 13.35562  5.493744 -0.07195938  11.56083  5.933572
#> 35   6.193479 -0.07626893 12.82779  5.447221 -0.06839260  11.42693  5.827882
#> 45   6.494207 -0.08099812 14.55438  5.460023 -0.06895616  12.35926  5.981840
#> 55   6.773051 -0.08148035 17.26816  5.335121 -0.06408350  13.57675  6.052669
#> 65   6.542885 -0.08148223 19.37808  4.804261 -0.05983977  14.03156  5.647593
#> 75   5.447844 -0.07205417 18.77615  3.755195 -0.04962356  12.66664  4.537017
#> all  8.716859 -0.13092988 22.68580  6.886534 -0.10034927  17.30197  7.817860
#>          all_lo   all_up
#> 5   -0.26785219 61.23686
#> 15  -0.11481287 17.90675
#> 25  -0.07779028 12.49425
#> 35  -0.07243817 12.15624
#> 45  -0.07506993 13.48213
#> 55  -0.07278725 15.43060
#> 65  -0.07033475 16.64109
#> 75  -0.05993572 15.51107
#> all -0.11599384 20.01740
#> 
#> $ear
#>         male   male_lo   male_up   female female_lo female_up      all
#> 5   8.834954 1.2943698 18.138433 5.799458 0.9404535 12.598648 7.440533
#> 15  6.663402 0.8638345 12.621382 4.432755 0.6277962  9.038465 5.663187
#> 25  6.125923 0.7928422 11.447897 4.170030 0.5855046  8.271526 5.244715
#> 35  5.931001 0.7705566 11.239644 4.130497 0.5689665  8.295145 5.106404
#> 45  5.652959 0.7552172 11.320950 4.066679 0.5799179  8.417920 4.910999
#> 55  5.116392 0.6632611 11.147375 3.842964 0.5843963  8.538808 4.517777
#> 65  4.333074 0.5675048 10.409819 3.377018 0.5188263  8.346105 3.871325
#> 75  3.254100 0.4262955  8.701093 2.629283 0.4157091  7.254095 2.944421
#> all 6.304237 0.8756158 11.957084 4.336539 0.6495997  8.688136 5.382736
#>        all_lo    all_up
#> 5   1.1314973 14.957146
#> 15  0.7717707 10.527392
#> 25  0.7073357  9.607655
#> 35  0.6945988  9.450180
#> 45  0.6933696  9.607121
#> 55  0.6429035  9.547658
#> 65  0.5695297  9.115761
#> 75  0.4257905  7.813362
#> all 0.7618394 10.054925
#>