Calculating population-averaged lifetime attributable risks

Calculate the population-averaged lifetime attributable risk due to radiation exposure.

Usage

population_LAR(
  dsGy,
  reference,
  riskmodel,
  agex = 1:8 * 10 - 5,
  PER = 10000,
  nmc = 10000
)

Arguments

dsGy

radiation dose in Gy or Sv (a single value).

reference

baseline rate, all cause mortality rate and age distribution in the reference population (a list object, which contains data.frame objects named baseline for baseline rates of the target endpoint, 'mortality' for all cause mortality rates and 'agedist' for age distribution in the reference population).

riskmodel

risk model risk model (a list object, which contains two list objects for excess relative risk model (err) and excess absolute risk model (ear), each of which contains a vector of parameter values (para), a matrix of variance covariance matrix (var), and a function to compute the risk given a parameter vector, a dose value, an age at exposure, an attained age and sex.

agex

a vector of ages at exposure, which represent age categories (default values: 5, 15, 25, ..., 75 to represent age categories 0-10, 10-20, ..., 70-80) option for risk calculation (a list object, which contains maximum age to follow up (an integer value)

PER

an integer (default value: 10000 to show the estimated risk as cases per 10000)

mmc

an integer for the Monte Carlo sample size (default: 10000)

Value

estimated risk information (list)

See also

Examples

 # The following examples use default data provided in CanEpiRisk package
 # for riskmodels (LSS_mortality and LSS_incidence) derived from Life Span Study
 # and baseline mortality and incidence rates for WHO global regions (Mortality and Incidence).

 # Example 1: allsolid mortality, Region-1, exposed to 0.1 Gy, followed up to age 100, LSS linear ERR
 ref1 <- list(  baseline=Mortality[[1]]$allsolid,     # baseline rates
               mortality=Mortality[[1]]$allcause,     # allcause mortality
                 agedist=agedist_rgn[[1]] )           # age distribution
 mod1 <- LSS_mortality$allsolid$L                     # risk model
 population_LAR( dsGy=0.1, reference=ref1, riskmodel=mod1 )    # CER cases per 10,000
#> $err
#>           male   male_lo   male_up     female  female_lo female_up        all
#> 5   130.553210 83.669992 203.53020 221.859630 148.199453 337.63506 175.293433
#> 15   92.224729 63.340254 132.86369 156.543696 115.687449 212.86277 123.759990
#> 25   65.630106 45.362486  91.83968 110.634928  86.012438 142.74316  87.816873
#> 35   46.604992 30.576172  68.48381  77.056894  57.842989 103.47762  61.903106
#> 45   32.688985 19.571985  53.35096  52.252585  35.523788  78.01505  42.699893
#> 55   21.299136 11.463337  39.21205  33.236911  20.018243  56.43535  27.530732
#> 65   12.247601  5.826844  25.87324  18.865345   9.934481  36.53418  15.840194
#> 75    5.789646  2.409148  13.96193   8.544794   3.938476  19.03778   7.378885
#> all  53.015810 36.595005  74.52241  83.906822  64.460820 109.31421  68.820120
#>         all_lo    all_up
#> 5   120.895894 260.03474
#> 15   95.430913 164.16882
#> 25   70.491257 110.83335
#> 35   47.123576  82.17626
#> 45   29.110157  63.68299
#> 55   16.470183  47.04022
#> 65    8.280112  31.18500
#> 75    3.350945  16.60112
#> all  54.786941  87.64512
#> 
#> $ear
#>          male   male_lo   male_up    female  female_lo female_up       all
#> 5   118.81129 77.066117 184.12852 177.96642 125.009162 256.20321 148.07158
#> 15   96.26584 65.613459 138.77072 143.51112 111.225505 187.28546 119.47647
#> 25   77.69444 53.398838 108.98101 115.27958  92.499141 144.29865  96.23987
#> 35   61.60623 40.757575  89.39057  91.42287  70.173648 118.10055  76.48642
#> 45   47.05154 29.014715  74.58106  70.12426  49.142966  99.74466  58.75787
#> 55   33.73453 19.012654  59.06217  50.82666  31.828938  80.48230  42.57951
#> 65   22.08762 11.254362  43.19379  33.38151  18.557958  59.48898  28.16513
#> 75   11.97137  5.463837  26.32955  18.13849   8.944523  36.50114  15.45929
#> all  61.39424 41.833103  87.00299  87.83595  69.697428 110.91000  74.86479
#>         all_lo    all_up
#> 5   106.254162 211.90441
#> 15   94.104662 155.01204
#> 25   78.027619 119.79718
#> 35   59.014279  99.44767
#> 45   41.238155  84.10431
#> 55   26.618488  68.22889
#> 65   15.677176  51.19411
#> 75    7.659102  31.93051
#> all  60.096746  94.40499
#> 

 # Example 2: leukaemia incidence, Region-4, exposed to 0.1 Gy, followed up to age 100, LSS LQ ERR
 ref2 <- list(  baseline=Incidence[[4]]$leukaemia,    # baseline rates
               mortality=Mortality[[4]]$allcause,     # all-cause mortality
                 agedist=agedist_rgn[[4]] )           # age distribution
 mod2 <- LSS_incidence$leukaemia$LQ                   # risk model
 population_LAR( dsGy=0.1, reference=ref2, riskmodel=mod2 )    # CER cases per 10,000
#> $err
#>          male    male_lo  male_up    female  female_lo female_up       all
#> 5   16.659787 0.26075176 69.78469 12.652240 0.20008845  51.59372 14.730493
#> 15   8.002893 0.11932843 19.34024  6.455588 0.09376754  15.18652  7.254996
#> 25   6.276006 0.08958518 13.11398  5.429399 0.07755391  11.35212  5.869386
#> 35   6.097068 0.09027265 12.73370  5.363137 0.07981379  11.29217  5.741280
#> 45   6.407269 0.09568479 14.54643  5.394629 0.08077693  12.34494  5.911646
#> 55   6.695994 0.10959690 17.52867  5.277803 0.08606164  13.78045  5.983616
#> 65   6.502435 0.11527993 19.66968  4.774096 0.08488327  14.27715  5.612856
#> 75   5.440662 0.09219976 19.04425  3.744906 0.06390690  12.76511  4.522909
#> all  8.666016 0.13795698 22.00663  6.856101 0.10754653  16.92812  7.772653
#>         all_lo   all_up
#> 5   0.23449004 61.11078
#> 15  0.10714942 17.34922
#> 25  0.08380734 12.27069
#> 35  0.08518026 12.03376
#> 45  0.08834573 13.46220
#> 55  0.09783648 15.65142
#> 65  0.09962338 16.89776
#> 75  0.07691414 15.61730
#> all 0.12305468 19.47840
#> 
#> $ear
#>         male   male_lo   male_up   female female_lo female_up      all
#> 5   8.859260 1.2668954 18.195974 5.794093 0.9218176 12.466334 7.487747
#> 15  6.684684 0.8757710 12.751435 4.428108 0.6448278  8.956400 5.687652
#> 25  6.191628 0.7891820 11.468337 4.154477 0.6084062  8.237021 5.289199
#> 35  5.980872 0.7925663 11.194832 4.101366 0.6033058  8.166290 5.119893
#> 45  5.701798 0.7634326 11.134395 4.046112 0.5998124  8.428402 4.938419
#> 55  5.191650 0.7063566 10.912272 3.825501 0.5817927  8.615603 4.557878
#> 65  4.392407 0.6126706 10.297003 3.364601 0.5123838  8.365563 3.896910
#> 75  3.283007 0.4484422  8.660284 2.627194 0.4132023  7.333779 2.958152
#> all 6.355054 0.8724096 11.806350 4.309807 0.6573355  8.644485 5.415701
#>        all_lo    all_up
#> 5   1.0900527 14.996870
#> 15  0.7614227 10.604514
#> 25  0.7096018  9.596534
#> 35  0.7017442  9.381993
#> 45  0.7084723  9.522952
#> 55  0.6544624  9.548709
#> 65  0.5679454  9.116486
#> 75  0.4313919  7.899438
#> all 0.7601004  9.967726
#>