Calculating population-averaged lifetime attributable risks

Calculate the population-averaged lifetime attributable risk due to radiation exposure.

Usage

population_LAR(
  dsGy,
  reference,
  riskmodel,
  agex = 1:8 * 10 - 5,
  PER = 10000,
  nmc = 10000
)

Arguments

dsGy

radiation dose in Gy or Sv (a single value).

reference

baseline rate, all cause mortality rate and age distribution in the reference population (a list object, which contains data.frame objects named 'baseline' for baseline rates of the target endpoint, 'mortality' for all cause mortality rates and 'agedist' for age distribution in the reference population).

riskmodel

risk model risk model (a list object, which contains two list objects for excess relative risk model (err) and excess absolute risk model (ear), each of which contains a vector of parameter values (para), a matrix of variance covariance matrix (var), and a function to compute the risk given a parameter vector, a dose value, an age at exposure, an attained age and sex.

agex

a vector of ages at exposure, which represent age categories (default values: 5, 15, 25, ..., 75 to represent age categories 0-10, 10-20, ..., 70-80) option for risk calculation (a list object, which contains maximum age to follow up (an integer value)

PER

an integer (default value: 10000 to show the estimated risk as cases per 10000)

mmc

an integer for the Monte Carlo sample size (default: 10000)

Value

estimated risk information (list)

See also

Examples

 # The following examples use default data provided in CanEpiRisk package
 # for riskmodels (LSS_mortality and LSS_incidence) derived from Life Span Study
 # and baseline mortality and incidence rates for WHO global regions (Mortality and Incidence).

 # Example 1: allsolid mortality, Region-1, exposed to 0.1 Gy, followed up to age 100, LSS linear ERR
 ref1 <- list(  baseline=Mortality[[1]]$allsolid,     # baseline rates
               mortality=Mortality[[1]]$allcause,     # allcause mortality
                 agedist=agedist_rgn[[1]] )           # age distribution
 mod1 <- LSS_mortality$allsolid$L                     # risk model
 population_LAR( dsGy=0.1, reference=ref1, riskmodel=mod1 )    # CER cases per 10,000
#> $err
#>           male   male_lo   male_up     female  female_lo female_up        all
#> 5   130.525587 82.492869 203.80023 221.864005 147.165583 340.05010 175.235394
#> 15   92.268422 62.457680 131.47600 156.174807 115.892832 215.32008 123.679918
#> 25   65.491978 44.687867  91.17580 110.307343  85.440657 143.49754  87.379861
#> 35   46.432053 30.279258  67.98389  76.739122  57.201921 103.14020  61.631180
#> 45   32.370714 19.504043  53.14938  51.999237  34.905138  77.69167  42.373855
#> 55   21.082226 11.406016  39.10720  33.041741  19.598072  55.86055  27.297865
#> 65   12.125670  5.788284  25.60954  18.739961   9.691158  36.32683  15.696646
#> 75    5.735749  2.404166  13.85522   8.500233   3.829937  18.92406   7.302441
#> all  52.918517 36.125116  73.92373  83.595727  64.294743 110.02254  68.529665
#>         all_lo    all_up
#> 5   120.397360 261.34958
#> 15   94.329103 164.93942
#> 25   69.688907 110.44664
#> 35   46.467490  81.52272
#> 45   28.730694  62.97902
#> 55   16.179715  46.25792
#> 65    8.153559  30.50437
#> 75    3.303835  16.22673
#> all  54.248850  87.39363
#> 
#> $ear
#>          male   male_lo   male_up    female  female_lo female_up       all
#> 5   118.89609 77.029205 185.64906 177.52319 125.561160 255.82002 147.62543
#> 15   96.21134 65.846781 140.03217 143.29420 110.862513 187.33380 119.31727
#> 25   77.66158 53.552913 109.17732 115.07012  91.969881 144.22214  96.10085
#> 35   61.45854 40.886804  89.29440  91.19958  69.800147 118.76854  76.40181
#> 45   46.89411 29.118085  73.47549  69.76484  48.785385  99.45969  58.67146
#> 55   33.57563 19.007949  57.95510  50.61073  31.689868  80.42400  42.57287
#> 65   21.99386 11.187092  42.56070  33.24656  18.579938  59.23399  28.20007
#> 75   11.97122  5.465173  25.76231  18.07269   8.872232  36.29747  15.50539
#> all  61.43311 42.117391  87.15885  87.47359  69.400027 111.09189  74.65814
#>         all_lo    all_up
#> 5   105.985995 212.12672
#> 15   94.000266 155.44582
#> 25   77.947763 119.50723
#> 35   58.881468  99.07553
#> 45   41.081129  83.40133
#> 55   26.611155  67.68109
#> 65   15.619270  50.18963
#> 75    7.606852  31.09182
#> all  60.144319  93.96373
#> 

 # Example 2: leukaemia incidence, Region-4, exposed to 0.1 Gy, followed up to age 100, LSS LQ ERR
 ref2 <- list(  baseline=Incidence[[4]]$leukaemia,    # baseline rates
               mortality=Mortality[[4]]$allcause,     # all-cause mortality
                 agedist=agedist_rgn[[4]] )           # age distribution
 mod2 <- LSS_incidence$leukaemia$LQ                   # risk model
 population_LAR( dsGy=0.1, reference=ref2, riskmodel=mod2 )    # CER cases per 10,000
#> $err
#>          male    male_lo  male_up    female  female_lo female_up       all
#> 5   17.039411 0.04473521 73.37549 12.924698 0.03401314  54.12304 15.072200
#> 15   8.113045 0.02289033 19.82155  6.534861 0.01855473  15.58163  7.358693
#> 25   6.354115 0.01864356 13.30080  5.500378 0.01614296  11.51166  5.943338
#> 35   6.187682 0.01852126 13.01728  5.444721 0.01624253  11.56191  5.825793
#> 45   6.496003 0.01973500 14.72372  5.462992 0.01699735  12.50541  5.990023
#> 55   6.773633 0.02648000 17.61870  5.336628 0.02083705  13.85354  6.053630
#> 65   6.613126 0.03334330 19.87546  4.837140 0.02384538  14.35952  5.695865
#> 75   5.541984 0.03348100 19.43300  3.815407 0.02269423  13.07369  4.603261
#> all  8.793231 0.02409633 22.88886  6.956704 0.01916309  17.38614  7.904535
#>         all_lo   all_up
#> 5   0.03960861 64.16881
#> 15  0.02082454 17.76009
#> 25  0.01744318 12.43861
#> 35  0.01741175 12.29398
#> 45  0.01833356 13.63621
#> 55  0.02366025 15.73712
#> 65  0.02845116 17.05342
#> 75  0.02765330 15.99280
#> all 0.02168202 20.11859
#> 
#> $ear
#>         male   male_lo   male_up   female female_lo female_up      all
#> 5   8.955387 1.1478087 18.173474 5.849873 0.8832394 12.538488 7.578118
#> 15  6.795092 0.8110560 12.683834 4.466869 0.6381805  8.851402 5.748863
#> 25  6.291253 0.7486772 11.303438 4.220957 0.5877802  8.185563 5.362239
#> 35  6.089641 0.7375452 11.124020 4.172044 0.5724519  8.189939 5.222116
#> 45  5.810406 0.7258781 11.154490 4.096192 0.5766877  8.384111 5.018011
#> 55  5.269103 0.6710254 10.941217 3.892891 0.5609549  8.465797 4.598279
#> 65  4.432730 0.5818644 10.301737 3.426684 0.4991856  8.253703 3.952021
#> 75  3.320127 0.4380317  8.696696 2.666833 0.3829109  7.272272 2.996503
#> all 6.452749 0.8019031 11.775187 4.369732 0.6342616  8.640441 5.514377
#>        all_lo    all_up
#> 5   1.0390744 15.090990
#> 15  0.7399892 10.411678
#> 25  0.6701712  9.470276
#> 35  0.6445767  9.258963
#> 45  0.6472412  9.456838
#> 55  0.6146849  9.449182
#> 65  0.5467926  9.028516
#> 75  0.4177396  7.834769
#> all 0.7343733  9.838162
#>